User can create/persist new logbooks and name them
MultiLogbookView receieves a user preferences storage object that they
write to with newly produced logs (via the blockchain interface) and
allows the user to long press the logbook cells and rename them. The
names persist strictly on the client side and the logbooks themselves
have randomly generated names. When we transition away from atra, based
on the architecture we're plannning, the logbooks will be blockchain
addresses from an HD wallet.