Setup framework for updating logs
Seems like we'll need a new database to keep track of which logs we want
updated post edit. We could have another database for the edited logs or
just use something like user preferences for a list of logs to update
when we come on line post editing them. The multi-select of logs is
working nicely.