User can edit single logs and re-log them
Logs have root transaction hashes that point to the original (ROOT) log.
They have current transaction hashes that distinguish them from the ROOT
log. I.E. all logs with current transaction hashes that differ from the
ROOT are children, or closed to the HEAD log.

To find the head log you can check the log's block time or get the image
record whose current multi hash matches the log's multi hash and order
the image records by timestamp.

I added a new data structure called the LogbookEntry which makes dealing
with logs and image records easier. You give it the total array of logs
and all the image records that reference the ROOT log (via
rootMultiHash) and the LogbookEntry will pick out all of the logs that
are children of the ROOT log and the order its ImageRecords
appropriately. It has a Log and ImageRecord property that will give you
the HEAD version of the Log and ImageRecord.