Add class to help decrypt and encrypt log metadata
Now we can simply instantiate a LogMetadata object with the appropriate
json data and keys and it will parse everything out for us. The detail
log view does this nicely. The single logbook view just parses images
and unless no image is present, in which case it decrypts logs and gets
their date stamp. TODO: we could probably just get the block time so as
to not slow down the loading of the logs with all of the necessary
decryptions.