Store private key in Keychain
On the native side we generate a 2048bit RSA key pair upon the first
login post installation and store the private key as the password for a
keychain entry. It's encrypted and requires the user to unlock the phone
to use it. The key generation takes a while but it only happens once per
installation.