Corroborate images with matching hashes
Now we're included a signed hash in the log metadata to prove we own the
public key listed as the key in the metadata dictionary. This allows for
corroborators to submit corroborating logs to the logbook with their own
public key and signature. Once we integrate ethers signing we can do
this all in one transaction (a txn from the corroborating wallet to the
logbook wallet).