Allow upload files to corroborate and check status
Now we can upload files to check if they're on chain and pass them on to
a single logbook viewer page that organizes them by logbook and their
log status.

Had to switch to a section list that contains a flatlist for this to
look nice but the section list is quite a bit slower than the flatlist,
or at least the combo is slower. TBD if we'll have time to fix up this
slow UX.

Had to change around the data structure that holds the logbook entries
to include sections, so now I just fixed the search function and am now
going to work on organizing by date + logbook.