Visualize logs on chain and those off together
The user can select multiple logs and can see when they were logged and
if they are not present on chain they'll have an orange border. TODO is
figure out how to add visualizations of how many times the image was
corroborated by other members. Perhaps we add in the evernym integration
so the user can see whose public keys audited the image?