Checking logs and ordering by blocktime
Now when we get logs from the chain we fill their blocktime property and
can use this to find the original log along the trunk of the log DAG as
well as the first log in the logbook to identify the owner of the log,
i.e. producer of root logs. From there we can find the tree and build
out branches of corroborating logs.